move pre_annotation scripts from notebooks folder to dcde
Scripts that had been in notebooks/pre_annotation folder were moved into dcde folder,
replacing old versions when necessary. Changed import statements as needed
